Crux.Gateway (crux_gateway v0.3.0-dev)

Main entry point for Crux.Gateway.

This modules fits under a supervision tree, see start_link/1 for configuration options.

Types

event()

Specs

event() :: {type :: atom(), data :: term(), shard_tuple()}

Type for events published by Crux.Gateway's shard gen stage stages.

For example: {:MESSAGE_CREATE, %{...}, {0, 1}}

gateway()
(since 0.2.0)

Specs

gateway() :: atom()

The given name for a Crux.Gateway process.

opts()
(since 0.3.0)

Specs

opts() :: opts_map() | (() -> opts_map())

Required to start Crux.Gateway.

These can optionally be a function for lazy loading, said function is applied exactly once with no arguments when the process is started.

Notes

:dispatcher - An atom representing a dispatcher module or a tuple of one and inital options.

Defaults to GenStage.DemandDispatcher.

:max_concurrency - Obtained from /gateway/bot (See also c:Crux.Rest.get_gateway_bot/0)

Defaults to 1

:presence - Used as initial presence for every session. (See also presence/0)

Defaults to none (the bot will be online with no activity)

:intents - The types of events you would like to receive from Discord. (See also Crux.Structs.Intents)

If none are provided here, you must provide them when starting a shard.

:name - Which name to use for this Crux.Gateway

:token - The bot token to use, you can get the one of your bot from here.

:url - Obtained from /gateway/bot (See also c:Crux.Rest.get_gateway_bot/0)

opts_map()
(since 0.3.0)

Specs

opts_map() ::
  %{
    optional(:dispatcher) =>
      module() | {module(), GenStage.Dispatcher.options()},
    optional(:max_concurrency) => pos_integer(),
    optional(:presence) => presence(),
    optional(:intents) => non_neg_integer(),
    :name => gateway(),
    :token => String.t(),
    :url => String.t()
  }
  | keyword()

See opts/0.

presence()
(since 0.2.0)

Specs

presence() :: presence_map() | (shard_id(), shard_count() -> presence_map())

Used as initial presence for every session. Can be either a map representing the presence to use or a function taking the shard id and the shard count and returning a presence map to use.

presence_map()
(since 0.3.0)

Specs

presence_map() :: %{
  optional(:status) => String.t(),
  optional(:activites) => [Crux.Gateway.Command.activity()]
}

See presence/0.

shard_count()
(since 0.3.0)

Specs

shard_count() :: pos_integer()

The shard count of a shard, shards are identified by a tuple of shard id and shard count.

shard_id()
(since 0.3.0)

Specs

shard_id() :: non_neg_integer()

The id of a shard, shards are identified by a tuple of shard id and shard count.

shard_opts()
(since 0.3.0)

Specs

shard_opts() :: %{
  optional(:intents) => non_neg_integer(),
  optional(:presence) => presence(),
  optional(:session_id) => String.t(),
  optional(:seq) => pos_integer(),
  :shard_id => shard_id(),
  :shard_count => shard_count()
}

Used to start a shard.

Notes:

:intents - What events you would like to receive from Discord. (See also Crux.Structs.Intents)

Optional if provided when starting Crux.Gateway, otherwise required.

If present these will override those provided when starting Crux.Gateway.

:presence - Used as initial presence for every session. (See also presence/0)

If present this will override the presenve provided when starting Crux.Gateway.

:session_id - If you want to (try to) resume a disconnected session, this also requires :seq to be set.

:seq - If you want to (try to) resume a disconnected session, this also requires :session_id to be set.

:shard_id - The id of the shard you want to start.

shard_count - The shard count of the shard you want to start.

shard_tuple()
(since 0.3.0)

Specs

shard_tuple() :: {shard_id(), shard_count()}

A tuple of shard id and shard count used to identify a shard.

Functions

child_spec(init_arg)

Specs

child_spec(opts()) :: Supervisor.child_spec()

Returns a specification to start this module under a supervisor.

See Supervisor.

producers(name)
(since 0.3.0)

Specs

producers(gateway()) :: %{required(shard_tuple()) => pid()}

A map of all running producers.

See event/0 for the type of published event.

send_command(gateway, shard_tuple, packet)
(since 0.3.0)

Specs

send_command(gateway(), shard_tuple(), Crux.Gateway.Command.command()) ::
  :ok | {:error, :not_found | :no_rate_limiter | :not_ready | :too_large}

Send a command to Discord through the specified shard.

Errors:

:not_found - The connection process for the given shard tuple wasn't found. (That shard likely wasn't started, or happened to just crash.)

:no_rate_limiter - The connection was found, but its rate limiter wasn't. (This means it crashed, that shouldn't happen.)

:not_ready - The connection is currently not ready, try again later.

:too_large - The given command was too large. (You shouldn't exceed this, the payload likely is invalid anyway.)

shards(name)
(since 0.3.0)

Specs

shards(gateway()) :: %{required(shard_tuple()) => pid()}

A map of all running shard supervisors.

start_link(opts)

Specs

start_link(opts()) :: Supervisor.on_start()

Start this module linked to the current process, intended to be used in combination with a Supervisor.

start_shard(name, shard)
(since 0.3.0)

Specs

start_shard(gateway(), shard_tuple() | shard_opts()) ::
  Supervisor.on_start_child()

Start a shard.

stop_shard(name, shard)
(since 0.3.0)

Specs

stop_shard(gateway(), shard_tuple()) :: :ok | {:error, :not_found}

Stop a shard.