Crux.Gateway.Command (crux_gateway v0.3.0-dev)

Builds Gateway Commands. Note: Only the sent ones can be found here.

A list of available op codes:

OP Code	Name	Direction

0	dispatch	received only

1	heartbeat	two way

2	identify	sent only

3	status_update	sent only

4	voice_state_update	sent only

5

6	resume	sent only

7	reconnect	received only

8	request_guild_members	sent only

9	invalid_session	received only

10	hello	received only

11	heartbeat_ack	received only

Gateway Lifecycle Flowchart

Types

activity()

Specs

activity() :: %{
  :name => String.t(),
  :type => non_neg_integer(),
  optional(:url) => String.t()
}

Used to set an activity via update_status/2.

:type must be a valid Activity Type

Note that streaming requires a twitch url pointing to a possible channel!

command()
(opaque)

Specs

command()

Encoded command ready to be sent to the gateway via Crux.Gateway.send_command/3.

If you want to build custom commands (read: new commands not yet supported by crux_gateway), build a valid Gateway Payload Structure using string keys(!) and pass it to encode_command/1.

Functions

encode_command(command)

Specs

encode_command(map()) :: command()

Encodes the given command map to a term that can be sent using Crux.Gateway.send_command/3.

heartbeat(sequence)

Specs

heartbeat(sequence :: non_neg_integer() | nil) :: command()

Builds a Heartbeat command.

Used to signalize the server that the client is still alive and able to receive messages.

Internally handled by Crux.Gateway already.

identify(args)

Specs

identify(
  data :: %{
    :shard_id => non_neg_integer(),
    :shard_count => pos_integer(),
    :token => String.t(),
    :intents => non_neg_integer(),
    optional(:presence) => Crux.Gateway.presence()
  }
) :: command()

Builds an Identify command.

Used to identify the gateway connection and "log in".

Internally handled by Crux.Gateway already.

request_guild_members(guild_id, opts \\ %{})

Specs

request_guild_members(
  guild_id :: Crux.Structs.Snowflake.t(),
  opts ::
    [
      query: String.t(),
      limit: non_neg_integer(),
      presences: boolean(),
      user_ids: Crux.Structs.Snowflake.t() | [Crux.Structs.Snowflake.t()],
      nonce: String.t()
    ]
    | map()
) :: command()

Builds a Request Guild Members command.

Used to request guild member for a specific guild.

Note: This must be sent to the connection handling the guild, otherwise the request will just be ignored.

The gateway will respond with :GUILD_MEMBERS_CHUNK packets until all (requested) members were received.

resume(map)

Specs

resume(
  data :: %{seq: non_neg_integer(), token: String.t(), session_id: String.t()}
) :: command()

Builds a Resume command.

Used to resume into a session which was unexpectly disconnected and may be resumable.

Internally handled by Crux.Gateway already.

update_status(status, activities \\ [])

Specs

update_status(status :: String.t(), activities :: [activity()] | []) ::
  command()

Builds a Update Status command.

Used to update the status of the client, including activity.

update_voice_state(guild_id, channel_id \\ nil, states \\ [])

Specs

update_voice_state(
  guild_id :: Crux.Structs.Snowflake.t(),
  channel_id :: Crux.Structs.Snowflake.t() | nil,
  states :: [self_mute: boolean(), self_deaf: boolean()]
) :: command()

Builds a Update Voice State command.

Used to join, switch between, and leave voice channels and/or change self_mute or self_deaf states.